+++ app/src/main/java/com/techcourse/controller/DefaultPathController.java
@@ -5,14 +5,14 @@ import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import web.org.springframework.web.bind.annotation.RequestMapping;
 import web.org.springframework.web.bind.annotation.RequestMethod;
-import webmvc.org.springframework.web.servlet.ModelAndView;
 import webmvc.org.springframework.web.servlet.view.JspView;
+import webmvc.org.springframework.web.servlet.view.ModelAndView;
 
 @Controller
 public class DefaultPathController {
 
     @RequestMapping(value = "/", method = RequestMethod.GET)
-    public ModelAndView getDefaultView(HttpServletRequest request, HttpServletResponse response){
+    public ModelAndView getDefaultView(HttpServletRequest request, HttpServletResponse response) {
         return new ModelAndView(new JspView("/index.jsp"));
     }
 }
