+++ mvc/build.gradle
@@ -11,6 +11,7 @@ repositories {
 }
 
 dependencies {
+    implementation "org.apache.tomcat.embed:tomcat-embed-core:10.1.13"
     implementation "jakarta.servlet:jakarta.servlet-api:5.0.0"
     implementation "javax.servlet.jsp:javax.servlet.jsp-api:2.3.3"
     implementation "jakarta.servlet.jsp.jstl:jakarta.servlet.jsp.jstl-api:3.0.0"
