+++ mvc/src/main/java/webmvc/org/springframework/web/servlet/exception/UncheckedServletException.java
@@ -1,4 +1,4 @@
-package com.techcourse;
+package webmvc.org.springframework.web.servlet.exception;
 
 public class UncheckedServletException extends RuntimeException {
