+++ mvc/src/main/java/webmvc/org/springframework/web/servlet/mvc/adaptor/AnnotationHandlerAdaptor.java
@@ -2,7 +2,7 @@ package webmvc.org.springframework.web.servlet.mvc.adaptor;
 
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
-import webmvc.org.springframework.web.servlet.ModelAndView;
+import webmvc.org.springframework.web.servlet.view.ModelAndView;
 import webmvc.org.springframework.web.servlet.mvc.tobe.HandlerExecution;
 
 public class AnnotationHandlerAdaptor implements HandlerAdaptor {
