+++ mvc/src/main/java/webmvc/org/springframework/web/servlet/mvc/adaptor/HandlerAdaptor.java
@@ -2,7 +2,7 @@ package webmvc.org.springframework.web.servlet.mvc.adaptor;
 
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
-import webmvc.org.springframework.web.servlet.ModelAndView;
+import webmvc.org.springframework.web.servlet.view.ModelAndView;
 
 public interface HandlerAdaptor {
