+++ mvc/src/main/java/webmvc/org/springframework/web/servlet/mvc/adaptor/HandlerAdaptors.java
@@ -4,7 +4,7 @@ import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import java.util.ArrayList;
 import java.util.List;
-import webmvc.org.springframework.web.servlet.ModelAndView;
+import webmvc.org.springframework.web.servlet.view.ModelAndView;
 
 public class HandlerAdaptors {
