+++ mvc/src/main/java/webmvc/org/springframework/web/servlet/view/JsonView.java
@@ -4,11 +4,7 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import java.io.Writer;
-import java.util.ArrayList;
-import java.util.Arrays;
-import java.util.List;
 import java.util.Map;
-import webmvc.org.springframework.web.servlet.View;
 
 public class JsonView implements View {
 
@@ -18,7 +14,7 @@ public class JsonView implements View {
     public void render(final Map<String, ?> model, final HttpServletRequest request, HttpServletResponse response)
             throws Exception {
         Writer writer = response.getWriter();
-        objectMapper.writeValue(writer,model);
+        objectMapper.writeValue(writer, model);
     }
 
     @Override
